requirements.txt
Django ==1.8.9
django-autoslug
pytz
